pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.75.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "routes>=2.5.1",
    "sendgrid>=6.11.0",
    "sqlalchemy>=2.0.40",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "oauthlib>=3.2.2",
    "wtforms>=3.2.1",
]
